Organizational changes:
* Improved Makefile
* Move source files to 'src/' and header files to 'inc/'
* Added a .gitignore
* Renamed 'Chanel' to 'Channel'
* Changed indentation to tabs (4 width)
* Added 42 headers to files
* Moved includes in '.cpp' files to their headers
* Added colors header
