+++ .gitignore
@@ -1,0 +1,3 @@
+.obj
+ircserv
+.vscode/
