+++ Makefile
@@ -1,18 +1,37 @@
-CXX = c++
-CXXFLAGS = -std=c++98 -Wall -Wextra -Werror
-SRC = main.cpp Server.cpp Client.cpp Chanel.cpp Commands.cpp
-OBJ = $(SRC:.cpp=.o)
 NAME = ircserv
+
+CXX = c++
+CXXFLAGS = -Wall -Werror -Wextra -std=c++98
+
+SRC_DIR = src
+INC_DIR = inc
+OBJ_DIR = .obj
+
+SRC_FILES = main.cpp Server.cpp Client.cpp Channel.cpp Commands.cpp
+INC_FILES = Server.hpp Client.hpp Channel.hpp Commands.hpp
+
+SRCS = $(addprefix $(SRC_DIR)/, $(SRC_FILES))
+HDRS = $(addprefix $(INC_DIR)/, $(INC_FILES))
+OBJS = $(addprefix $(OBJ_DIR)/, $(SRCS:%.cpp=%.o))
 
 all: $(NAME)
 
-$(NAME): $(OBJ)
-	$(CXX) $(CXXFLAGS) -o $(NAME) $(OBJ)
+$(NAME): $(OBJS) $(HDRS)
+	@$(CXX) $(CXXFLAGS) $(OBJS) -o $(NAME)
+	@echo "Compiled!"
+
+$(OBJ_DIR)/%.o: %.cpp $(HDRS)
+	@mkdir -p $(OBJ_DIR)/$(dir $<)
+	@$(CXX) $(CXXFLAGS) -c $< -o $@
 
 clean:
-	rm -f $(OBJ)
+	@rm -rf $(OBJS) $(OBJ_DIR)
+	@echo "Object files removed"
 
 fclean: clean
-	rm -f $(NAME)
+	@rm -rf $(NAME)
+	@echo "Executable removed"
 
 re: fclean all
+
+.PHONY: all clean fclean re val
