+++ inc/colors.hpp
@@ -1,0 +1,63 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   colors.hpp                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mortins- <mortins-@student.42lisboa.com    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/25 17:54:29 by mortins-          #+#    #+#             */
+/*   Updated: 2024/11/25 17:54:31 by mortins-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef COLORS_HPP
+# define COLORS_HPP
+
+// Styles
+# define RESET			"\e[0m"
+# define BOLD			"\e[1m"
+# define ITALIC			"\e[3m"
+# define UNDERLINE		"\e[4m"
+# define STRIKETHROUGH	"\e[9m"
+
+// Regular Colors
+# define BLACK	"\e[30m"
+# define RED	"\e[31m"
+# define GREEN	"\e[32m"
+# define YELLOW	"\e[33m"
+# define BLUE	"\e[34m"
+# define PURPLE	"\e[35m"
+# define CYAN	"\e[36m"
+# define WHITE	"\e[37m"
+
+// High Intensity Colors
+# define BLACK_H	"\e[90m"
+# define RED_H		"\e[91m"
+# define GREEN_H	"\e[92m"
+# define YELLOW_H	"\e[93m"
+# define BLUE_H		"\e[94m"
+# define PURPLE_H	"\e[95m"
+# define CYAN_H		"\e[96m"
+# define WHITE_H	"\e[97m"
+
+// Background
+# define BG_BLACK	"\e[40m"
+# define BG_RED		"\e[41m"
+# define BG_GREEN	"\e[42m"
+# define BG_YELLOW	"\e[43m"
+# define BG_BLUE	"\e[44m"
+# define BG_PURPLE	"\e[45m"
+# define BG_CYAN	"\e[46m"
+# define BG_WHITE	"\e[47m"
+
+// High Intensity Background
+# define BG_BLACK_H		"\e[100m"
+# define BG_RED_H		"\e[101m"
+# define BG_GREEN_H		"\e[102m"
+# define BG_YELLOW_H	"\e[103m"
+# define BG_BLUE_H		"\e[104m"
+# define BG_PURPLE_H	"\e[105m"
+# define BG_CYAN_H		"\e[106m"
+# define BG_WHITE_H		"\e[107m"
+
+#endif
